Extract a new package from the generic bits
