+++ pkg/language/language.go
@@ -1,23 +1,18 @@
 package language
 
 import (
-	"archive/tar"
-	"compress/gzip"
 	"context"
 	"errors"
 	"fmt"
-	"io"
-	"net/http"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"regexp"
 	"runtime"
 	"strings"
-	"time"
 
 	"github.com/anttiharju/vmatch/pkg/exitcode"
 	"github.com/anttiharju/vmatch/pkg/finder"
+	"github.com/anttiharju/vmatch/pkg/install"
 	"github.com/anttiharju/vmatch/pkg/wrapper"
 )
 
@@ -89,100 +84,23 @@ func (w *WrappedLanguage) noBinary() bool {
 }
 
 func (w *WrappedLanguage) install(ctx context.Context) {
-	goVersion := w.DesiredVersion
-	goOS := runtime.GOOS
-	goArch := runtime.GOARCH
-	installPath := w.InstallPath
+	downloadURL := buildDownloadURL(w.DesiredVersion)
 
-	// Ensure install directory exists
-	err := os.MkdirAll(installPath, 0o755)
+	err := install.FromURL(ctx, downloadURL, w.InstallPath)
 	if err != nil {
-		w.ExitWithPrint(exitcode.CmdStartIssue, "failed to create install directory: "+err.Error())
+		w.ExitWithPrint(exitcode.CmdStartIssue, "failed to install Go: "+err.Error())
 	}
 
-	// Download and extract Go
-	url := fmt.Sprintf("https://go.dev/dl/go%s.%s-%s.tar.gz", goVersion, goOS, goArch)
-
-	ctx, cancel := context.WithTimeout(ctx, 30*time.Second)
-	defer cancel()
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
-	if err != nil {
-		w.ExitWithPrint(exitcode.CmdStartIssue, "creating request: "+err.Error())
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		w.ExitWithPrint(exitcode.CmdStartIssue, "failed to download Go: "+err.Error())
-	}
-
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		msg := fmt.Sprintf("failed to download Go: received status code %d", resp.StatusCode)
-		w.ExitWithPrint(exitcode.CmdStartIssue, msg)
-	}
-
-	// Decompress and extract tar.gz
-	gzr, err := gzip.NewReader(resp.Body)
-	if err != nil {
-		w.ExitWithPrint(exitcode.CmdStartIssue, "failed to create gzip reader: "+err.Error())
-	}
-	defer gzr.Close()
-
-	tarReader := tar.NewReader(gzr)
-
-	// Extract files, stripping the top-level directory
-	for {
-		header, err := tarReader.Next()
-		if errors.Is(err, io.EOF) {
-			break
-		}
-
-		if err != nil {
-			w.ExitWithPrint(exitcode.CmdStartIssue, "error extracting tar archive: "+err.Error())
-		}
-
-		// Skip the top-level directory by removing the first path component
-		parts := strings.SplitN(header.Name, "/", 2)
-		if len(parts) < 2 {
-			continue // Skip if there's no second part (top-level directory itself)
-		}
-
-		target := filepath.Join(installPath, parts[1])
-
-		switch header.Typeflag {
-		case tar.TypeDir:
-			// Create directory
-			err = os.MkdirAll(target, 0o755)
-			if err != nil {
-				w.ExitWithPrint(exitcode.CmdStartIssue, "failed to create directory: "+err.Error())
-			}
-		case tar.TypeReg:
-			// Create file
-			dir := filepath.Dir(target)
-			if err := os.MkdirAll(dir, 0o755); err != nil {
-				w.ExitWithPrint(exitcode.CmdStartIssue, "failed to create file directory: "+err.Error())
-			}
-
-			f, err := os.OpenFile(target, os.O_CREATE|os.O_RDWR, os.FileMode(header.Mode))
-			if err != nil {
-				w.ExitWithPrint(exitcode.CmdStartIssue, "failed to create file: "+err.Error())
-			}
-
-			if _, err := io.Copy(f, tarReader); err != nil {
-				f.Close()
-				w.ExitWithPrint(exitcode.CmdStartIssue, "failed to write file: "+err.Error())
-			}
-
-			f.Close()
-		}
-	}
-
-	// Verify installation
 	if w.noBinary() {
 		w.ExitWithPrint(exitcode.CmdStartIssue, "failed to install Go: binary not found after installation")
 	}
+}
+
+func buildDownloadURL(version string) string {
+	return fmt.Sprintf("https://go.dev/dl/go%s.%s-%s.tar.gz",
+		version,
+		runtime.GOOS,
+		runtime.GOARCH)
 }
 
 func (w *WrappedLanguage) getGoPath() string {
